Refactor out global.service - create HttpUtilsService
