+++ src/client/app/shared/cors-site/cors-site.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { Http } from '@angular/http';
 import { Observable } from 'rxjs/Rx';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 import { GlobalService } from '../global/global.service';
 import { SelectSiteSearchType, WFSService } from '../wfs/wfs.service';
+import { HttpUtilsService } from '../global/http-utils.service';
 
 /**
  * This class provides the service with methods to retrieve CORS sites from DB and select site.
@@ -59,23 +60,13 @@ export class CorsSiteService {
    */
   getAllCorsSites(): Observable<any[]> {
     return this.http.get(this.globalService.getWebServiceURL()+'/corsSites?size=1000')
-      .map((response: Response) => response.json())
-      .catch(this.handleError);
+      .map(HttpUtilsService.handleJsonData)
+      .catch(HttpUtilsService.handleError);
   }
 
   getSiteById(id: number): Observable<any> {
     return this.http.get(this.globalService.getWebServiceURL() + '/corsSites?id=' + id)
-      .map((response: Response) => response.json())
-      .catch(this.handleError);
-  }
-
-  /**
-   * Handle HTTP error
-   */
-  private handleError(error: any) {
-    let errMsg = (error.message) ? error.message :
-      error.status ? `${error.status} - ${error.statusText}` : 'Server error';
-    console.error(errMsg);
-    return Observable.throw(errMsg);
+      .map(HttpUtilsService.handleJsonData)
+      .catch(HttpUtilsService.handleError);
   }
 }
