+++ src/client/app/shared/name-list/name-list.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { Http } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
+import { HttpUtilsService } from '../global/http-utils.service';
 
 /**
  * This class provides the NameList service with methods to read names and add names.
@@ -22,19 +23,7 @@ export class NameListService {
    */
   get(): Observable<string[]> {
     return this.http.get('/assets/data.json')
-            .map((res: Response) => res.json())
-            .catch(this.handleError);
-  }
-
-  /**
-    * Handle HTTP error
-    */
-  private handleError (error: any) {
-    // In a real world app, we might use a remote logging infrastructure
-    // We'd also dig deeper into the error to get a better message
-    let errMsg = (error.message) ? error.message :
-      error.status ? `${error.status} - ${error.statusText}` : 'Server error';
-    console.error(errMsg); // log to console instead
-    return Observable.throw(errMsg);
+            .map(HttpUtilsService.handleJsonData)
+            .catch(HttpUtilsService.handleError);
   }
 }
